Restructure Rust side, add sink
